requirements_2.txt
aiobotocore==2.21.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4

arrow==1.3.0

attrs==25.3.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
bleach==6.2.0
boltons==25.0.0
botocore==1.37.1
branca==0.8.1


build==1.2.2.post1
cachetools==5.5.2
# catboost==1.2.7
category_encoders==2.8.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.1
cmake==4.0.0
colorlog==6.9.0

compress-pickle==2.1.0
contourpy==1.3.1
cookiecutter==2.6.0
cramjam==2.9.1
cycler==0.12.1
cykhash==2.0.1
Cython==3.0.12


defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
dotenv==0.9.9
dynaconf==3.2.10
et_xmlfile==2.0.0


fastapi==0.115.12
fastjsonschema==2.21.1
fastparquet==2024.11.0
fiona==1.10.1
folium==0.19.5
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
graphviz==0.20.3
grpcio==1.71.0
h11==0.14.0

h3==4.3.1
h3-pyspark==1.2.6
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
implicit==0.7.2

importlib_resources==6.5.2




Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1


jupyterlab_pygments==0.3.0
kedro==0.19.12
kedro-datasets==6.0.0
kedro-telemetry==0.6.2
kiwisolver==1.4.8
lazy_loader==0.4
lightgbm==4.6.0
line_profiler==4.2.0
llvmlite==0.44.0
lz4==4.4.3
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1

mdurl==0.1.2
mistune==3.1.3
more-itertools==10.6.0
mpu==0.23.1
mrmr-selection==0.2.8
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4

ninja==1.11.1.4
numba==0.61.0
numexpr==2.10.2
numpy==1.26.4
omegaconf==2.3.0
openpyxl==3.1.5
optuna==4.2.1
osmium==4.0.2

pandas==2.2.3
pandera==0.23.1
pandocfilters==1.5.1
parse==1.20.2

patsy==1.0.1


pillow==11.1.0

plotly==6.0.1
pluggy==1.5.0
polars==1.27.1
pre_commit_hooks==5.0.0

propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2



py4j==0.10.9.7
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pyDOE2==1.3.0
pyDOE3==1.0.5

pymssql==2.3.2
pyodbc==5.2.0
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.1
pyproject_hooks==1.2.0
pyrobuf==0.9.3
pyrosm==0.6.2
pyspark==3.5.2

python-dotenv==1.1.0
python-rapidjson==1.20
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2025.2
PyYAML==6.0.2

referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rope==1.13.0
rpds-py==0.24.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3fs==2025.3.0
scikit-build==0.18.1
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
shap==0.47.1
shapely==2.0.7

slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40

starlette==0.46.1
statsmodels==0.14.4
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2

tqdm==4.67.1

typeguard==4.4.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0

tzdata==2025.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0

webencodings==0.5.1
wheel==0.45.1
wrapt==1.17.2
xgboost==3.0.0
xyzservices==2025.1.0
yarl==1.18.3

